+++ packages/orq-rc/RELEASES.md
@@ -2039,3 +2039,13 @@ Based on:
 - [typescript v3.8.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.8 - packages/orq-rc
+
+## 2025-06-10 06:27:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.558.1 (2.623.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.8.0-rc.9] packages/orq-rc
+### Releases
+- [NPM v3.8.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.9 - packages/orq-rc